No gene symbol can be found for any ID
When using "queryMany" to convert gene ID to gene symbol, if no symbol can be found for any gene ID, an error will be thrown out and the workflow will be stuck there. So this is to take care of that issue: the error will still be thrown out, but the workflow will continue (keeping the original gene ID without converting).